Add some features
