+++ packages/ogm/src/classes/Model.ts
@@ -22,7 +22,7 @@ import type { DocumentNode, GraphQLSchema, SelectionSetNode } from "graphql";
 import { graphql, parse, print } from "graphql";
 import { debuglog } from 'util';
 import type { DeleteInfo, GraphQLOptionsArg, GraphQLWhereArg } from "../types";
-import { buildGQLOptions } from "../utils/graphql-options";
+import generateSelectionSet from "../utils/generate-selection-set";
 
 type Neo4jGraphQLOGMContext = Omit<Neo4jGraphQLContext, "jwt" | "token">;
 
@@ -174,34 +174,7 @@ class Model {
         context?: any;
         rootValue?: any;
     } = {}): Promise<T> {
-        const recursive = (obj: Record<string, any>) => {
-            return Object.entries(obj)
-                .map(([k, v]) => {
-                    if (v === true) return k;
-                    const valueKeys = Object.keys(v);
-                    if (["where", "directed", "options", "after", "first", "sort"].some((v) => valueKeys.includes(v))) {
-                        const options = Object.entries(v)
-                            .filter(([k]) => ["where", "directed", "options", "after", "first", "sort"].includes(k))
-                            .reduce((acc, [k, v]) => ({ ...acc, [k]: v }), {});
-                        const { where: w, directed: d, options: o, after: a, first: f, sort: s, ...other } = v;
-                        return `${k} (${buildGQLOptions(options)}) {\n${recursive(other)}\n}`;
-                    }
-
-                    if (valueKeys.includes("on")) {
-                        const entries = Object.entries(v["on"])
-                            .map(([modelName, value]) => {
-                                return `... on ${modelName} {\n${recursive(value as Record<string, any>)}\n}`;
-                            })
-                            .join("\n");
-                        const { on, ...other } = v;
-                        return `${k} {\n${recursive(other)}\n${entries}\n}`;
-                    }
-
-                    return `${k} {\n${recursive(v)}\n}`;
-                })
-                .join("\n");
-        };
-        const stringifiedSelectionSet = recursive(selectionSet);
+        const stringifiedSelectionSet = generateSelectionSet(selectionSet);
         logger("FindSafe with Object Selection Set: %s and stringified %s", selectionSet, stringifiedSelectionSet);
         return this.find({ selectionSet: `{${stringifiedSelectionSet}}`, ...otherProps });
     }
