+++ packages/ogm/src/find-safe.ts
@@ -1,5 +1,5 @@
 const findSafeGenerated = `import type { RawGQL } from '@neo4j/graphql-ogm';
-
+export type RequiredResolvers = Required<Resolvers>;
 export type Primitive = string | number | boolean;
 export type Prettify<T> = {
     [K in keyof T]: T[K];
