+++ packages/ogm/src/index.ts
@@ -21,4 +21,5 @@ export * from "./classes";
 export { default as RawGQL } from './classes/RawGQL';
 export { default as generate } from "./generate";
 export { DeleteInfo, GraphQLOptionsArg, GraphQLSortArg, GraphQLWhereArg } from "./types";
+export { default as generateSelectionSet } from "./utils/generate-selection-set";
